requirements.txt
# Core Backend Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
python-dotenv>=1.0.0

# LangChain and LLM
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
# Optional: Ollama for local embeddings (uncomment if you want to use Ollama)
# langchain-ollama>=0.1.0

# OpenAI API (for ChatGPT and Whisper API)
openai>=1.0.0

# Google TTS
gTTS==2.5.0

# Audio Processing
pydub==0.25.1
SpeechRecognition==3.10.1
ffmpeg-python>=0.2.0

# CORS and WebSocket support
python-socketio==5.11.0
aiofiles==23.2.1
websockets>=12.0

# Utilities
pydantic>=2.10.0
pydantic-settings>=2.7.0

# MongoDB
pymongo>=4.6.0